parse: add mnl_attr_for_each_payload
This function allows you to parse the payload of the Netlink message.
Thus, you can skip the headers and go down to the part of the message
you want to parse.

This patch has updated the LIBVERSION and it also bumps the version
to 1.0.3, in case that someone wants to package some snapshot, not to
mess with the new symbol we have added.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>