libmnl: zero attribute padding
Sergei Trofimovich reports 'uninitialized bytes' warnings from nftables:

Syscall param sendmsg(msg.msg_iov[0]) points to uninitialised byte(s)
   at 0x55B9EFB: sendmsg (in /lib64/libc-2.25.so)
   by 0x43E658: mnl_nft_socket_sendmsg (mnl.c:239)
   by 0x43E658: mnl_batch_talk (mnl.c:254)
   by 0x407898: nft_netlink (libnftables.c:58)
   by 0x407898: nft_run (libnftables.c:96)
   by 0x407CD5: nft_run_cmd_from_buffer (libnftables.c:291)
   by 0x406EDE: main (main.c:274)

This is harmless, the uninitialized memory is the padding
that sometimes needs to be inserted between end of an attribute
and the beginning of the new attribute.

Zero it to silence memory sanitizer output.

Signed-off-by: Florian Westphal <fw@strlen.de>
Acked-by: Pablo Neira Ayuso <pablo@netfilter.org>